Author2

+++ sql/setup.sql
@@ -36,7 +36,7 @@ CREATE TABLE github_users (
   id BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
   username TEXT NOT NULL,
   email TEXT,
-  cohort_id INT,
+  cohort_id BIGINT,
   role TEXT,
   -- github TEXT
   FOREIGN KEY(cohort_id) REFERENCES cohorts(id)
